linux-user: Implement interrupting syscalls
When a syscall waits for something a signal handler is supposed
to do (e.g. reading from a pipe to be written when SIGCHLD arrives),
and that signal arrives during a restarting syscall (or between
checking for presence of signals and doing the host syscall) then
the target signal handler won't be run.  We use longjmp to
leave the critical section, deliver the signal and restart the
syscall.