linux-user: Don't wrap wait4
This points to a general problem with the siglongjmp approach
to exit the signal handler.  If the syscall already returned to
userspace, but the flag is still set (small but possible window),
we loose the return value of the syscall as we restart it.
For wait() it's worse, we loose the fact that there were childs
at all (the syscall returned the pid of one of them), and restarting
the wait() can exit with ECHILD.  Paper over the problem for wait.